feat: normalize moderator_warning type and reorganize components
- Standardize moderation notification type to 'moderator_warning'
- Add migration to update existing notifications and type constraint
- Move QuestHistoryCard from rows to cards folder for proper taxonomy
- Fix NotificationRow controlled/uncontrolled state handling
- Update all barrel exports to reflect structural changes
