+++ app/components/cards/index.ts
@@ -1,3 +1,4 @@
-export { default as QuestCard } from './QuestCard';
-export { default as ShopItemCard } from './ShopItemCard';
-export { default as CompactQuestCard } from './CompactQuestCard';
+export { default as QuestCard } from "./QuestCard";
+export { default as ShopItemCard } from "./ShopItemCard";
+export { default as CompactQuestCard } from "./CompactQuestCard";
+export { default as QuestHistoryCard } from "./QuestHistoryCard";
