+++ app/components/index.ts
@@ -1,32 +1,51 @@
 // Button Components
-export { Button, RatingButton, NotificationsButton, RatingButtonRow, CategorySelectButton, InlineCtaButton } from './buttons';
+export {
+    Button,
+    CategorySelectButton,
+    InlineCtaButton,
+    NotificationsButton,
+    RatingButton,
+    RatingButtonRow,
+} from "./buttons";
 
 // Chip Components
-export { EarnedChip, StatusPill, RarityBadge } from './chips';
+export { EarnedChip, RarityBadge, StatusPill } from "./chips";
 
 // Input Components
-export { TextInput, AuthTab, DropdownSelectField, PasswordStrengthIndicator } from './inputs';
+export {
+    AuthTab,
+    DropdownSelectField,
+    PasswordStrengthIndicator,
+    TextInput,
+} from "./inputs";
 
 // Icon Components
-export { RatingReceivedIcon } from './icons';
+export { RatingReceivedIcon } from "./icons";
 
 // Loading Components
-export { LoadingDots } from './loading';
+export { LoadingDots } from "./loading";
 
 // Avatar Components
-export { AvatarGridItem } from './avatars';
+export { AvatarGridItem } from "./avatars";
 
 // Filter Components
-export { CategoryFilterBar, FilterToggle, HistoryFilterPill, HistoryFilterRow, ShopFilterPill, CategoryButtonRow } from './filters';
+export {
+    CategoryButtonRow,
+    CategoryFilterBar,
+    FilterToggle,
+    HistoryFilterPill,
+    HistoryFilterRow,
+    ShopFilterPill,
+} from "./filters";
 
 // Navigation Components
-export { NavItem, BottomNavBar } from './navigation';
+export { BottomNavBar, NavItem } from "./navigation";
 
 // Row Components
-export { NotificationRow, CommentRow, ActiveQuestRow, QuestHistoryCard } from './rows';
+export { ActiveQuestRow, CommentRow, NotificationRow } from "./rows";
 
 // Card Components
-export { QuestCard, ShopItemCard } from './cards';
+export { QuestCard, QuestHistoryCard, ShopItemCard } from "./cards";
 
 // Badge Components
-export { BadgeSelectorItem } from './badges';
+export { BadgeSelectorItem } from "./badges";
