+++ app/components/rows/index.ts
@@ -1,5 +1,8 @@
-export { default as NotificationRow } from './NotificationRow';
-export type { NotificationRowProps, NotificationState, NotificationType } from './NotificationRow';
-export { default as CommentRow } from './CommentRow';
-export { default as ActiveQuestRow } from './ActiveQuestRow';
-export { default as QuestHistoryCard } from './QuestHistoryCard';
+export { default as NotificationRow } from "./NotificationRow";
+export type {
+    NotificationRowProps,
+    NotificationState,
+    NotificationType,
+} from "./NotificationRow";
+export { default as CommentRow } from "./CommentRow";
+export { default as ActiveQuestRow } from "./ActiveQuestRow";
